Improve file organization by breaking up into feature-based classes.
