+++ UnIt/Source/CGFloat+Precision.swift
@@ -1,6 +1,6 @@
 public extension CGFloat {
     func roundTo(decimalPlace: Int) -> CGFloat {
-        let divisor = CGFloat(10^decimalPlace)
+        let divisor = CGFloat(10 ^ decimalPlace)
         return (self * divisor).rounded() / divisor
     }
 }
